Moving some project files around
