+++ packages/cli/commands/create/app.js
@@ -1,8 +1,8 @@
-const { createProject } = require('@hubspot/cli-lib/projects');
+const { cloneGitHubRepo } = require('@hubspot/cli-lib/github');
 
 module.exports = {
   hidden: true,
   dest: ({ name, assetType }) => name || assetType,
   execute: async ({ options, dest, assetType }) =>
-    createProject(dest, assetType, 'crm-card-weather-app', '', options),
+    cloneGitHubRepo(dest, assetType, 'crm-card-weather-app', '', options),
 };
